Fix authentication and input validation security issues
- Add server-side auth checks (getServerSession + redirect) to 13 pages
  that were only protected client-side
- Refactor timer page to server component wrapper + client component
  to enable server-side auth check
- Add date format regex validation before $queryRaw in workout router
- Add .max() length limits to string inputs in workout, goal, and
  workoutMenu routers to prevent oversized payloads
- Remove tRPC path from error message in WorkoutAddPage to prevent
  API path information leakage
