+++ src/app/goal-edit/page.tsx
@@ -1,9 +1,15 @@
 import { Suspense } from "react";
+import { getServerSession } from "next-auth";
+import { redirect } from "next/navigation";
+import { authOptions } from "../../pages/api/auth/[...nextauth]";
 import { GoalEditPage } from "./GoalEditPage";
 import { Container, Heading, Navigation } from "../../components/server";
 import { Loading } from "../../components";
 
-export default function Page() {
+export default async function Page() {
+  const session = await getServerSession(authOptions);
+  if (!session?.user) redirect('/login');
+
   return (
     <>
       <main className="md:mt-4">
