+++ src/app/workout-history/page.tsx
@@ -1,7 +1,13 @@
+import { getServerSession } from "next-auth";
+import { redirect } from "next/navigation";
+import { authOptions } from "../../pages/api/auth/[...nextauth]";
 import { WorkoutHistoryPage } from "./WorkoutHistoryPage";
 import { Container, Heading, Navigation } from "../../components/server";
 
-export default function Page() {
+export default async function Page() {
+  const session = await getServerSession(authOptions);
+  if (!session?.user) redirect('/login');
+
   return (
     <>
       <main className="md:mt-4">
