+++ src/app/workout-recorder/page.tsx
@@ -1,8 +1,14 @@
+import { Suspense } from "react";
+import { getServerSession } from "next-auth";
+import { redirect } from "next/navigation";
+import { authOptions } from "../../pages/api/auth/[...nextauth]";
 import { WorkoutRecorderPage } from "./WorkoutRecorderPage";
 import { Container, Heading, Navigation } from "../../components/server";
-import { Suspense } from "react";
 
-export default function Page() {
+export default async function Page() {
+  const session = await getServerSession(authOptions);
+  if (!session?.user) redirect('/login');
+
   return (
     <>
       <main className="md:mt-4">
